+++ Mailer/extensions/mailer-module/src/migrations/20251216085341_mail_job_batch_status_index.php
@@ -1,0 +1,22 @@
+<?php
+
+declare(strict_types=1);
+
+use Phinx\Migration\AbstractMigration;
+
+final class MailJobBatchStatusIndex extends AbstractMigration
+{
+    public function up(): void
+    {
+        $this->table('mail_job_batch')
+            ->addIndex('status')
+            ->update();
+    }
+
+    public function down(): void
+    {
+        $this->table('mail_job_batch')
+            ->removeIndex('status')
+            ->update();
+    }
+}